Moved NewProjectNameModalProps and NewProjectNameModal to src/modal/NewProjectNameModal.tsx.
